+++ src/RcppExports.cpp
@@ -490,6 +490,37 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// cpp_cudnn_runtime_version
+int64_t cpp_cudnn_runtime_version();
+RcppExport SEXP _torch_cpp_cudnn_runtime_version() {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    rcpp_result_gen = Rcpp::wrap(cpp_cudnn_runtime_version());
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_cudnn_is_available
+bool cpp_cudnn_is_available();
+RcppExport SEXP _torch_cpp_cudnn_is_available() {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    rcpp_result_gen = Rcpp::wrap(cpp_cudnn_is_available());
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_cuda_memory_stats
+torch::vector::int64_t cpp_cuda_memory_stats(int64_t device);
+RcppExport SEXP _torch_cpp_cuda_memory_stats(SEXP deviceSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< int64_t >::type device(deviceSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_cuda_memory_stats(device));
+    return rcpp_result_gen;
+END_RCPP
+}
 // cpp_device_type_to_string
 std::string cpp_device_type_to_string(Rcpp::XPtr<XPtrTorchDevice> device);
 RcppExport SEXP _torch_cpp_device_type_to_string(SEXP deviceSEXP) {
@@ -37224,6 +37255,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_torch_cpp_cuda_device_count", (DL_FUNC) &_torch_cpp_cuda_device_count, 0},
     {"_torch_cpp_cuda_current_device", (DL_FUNC) &_torch_cpp_cuda_current_device, 0},
     {"_torch_cpp_cuda_get_device_capability", (DL_FUNC) &_torch_cpp_cuda_get_device_capability, 1},
+    {"_torch_cpp_cudnn_runtime_version", (DL_FUNC) &_torch_cpp_cudnn_runtime_version, 0},
+    {"_torch_cpp_cudnn_is_available", (DL_FUNC) &_torch_cpp_cudnn_is_available, 0},
+    {"_torch_cpp_cuda_memory_stats", (DL_FUNC) &_torch_cpp_cuda_memory_stats, 1},
     {"_torch_cpp_device_type_to_string", (DL_FUNC) &_torch_cpp_device_type_to_string, 1},
     {"_torch_cpp_device_index_to_int", (DL_FUNC) &_torch_cpp_device_index_to_int, 1},
     {"_torch_cpp_torch_device", (DL_FUNC) &_torch_cpp_torch_device, 2},