+++ src/cuda.cpp
@@ -13,3 +13,14 @@ int64_t cpp_cuda_current_device() { return lantern_cuda_current_device(); }
 XPtrTorchvector_int64_t cpp_cuda_get_device_capability(int64_t device) {
   return lantern_cuda_get_device_capability(device);
 }
+
+// [[Rcpp::export]]
+int64_t cpp_cudnn_runtime_version() { return lantern_cudnn_runtime_version(); }
+
+// [[Rcpp::export]]
+bool cpp_cudnn_is_available() { return lantern_cudnn_is_available(); }
+
+// [[Rcpp::export]]
+torch::vector::int64_t cpp_cuda_memory_stats(int64_t device) {
+  return torch::vector::int64_t(lantern_cuda_device_stats(device));
+}