+++ NEWS.md
@@ -18,6 +18,8 @@
 - Fixed a bug that caused `autograd_grad` to deadlock when used with custom autograd functions. (#771)
 - Fixed a bug in `torch_max` and `torch_min` that would fail with `length=2` Tensors. (#772)
 - Added `nn_flatten` and `nn_unflatten` modules. (#773)
+- Added `cuda_memory_stats()` and `cuda_memory_summary()` to verify the amount of memory torch is using from the GPU. (#774)
+- Added `backends_cudnn_version()` to query the CuDNN version found by torch. (#774)
 
 # torch 0.6.1